refactor: integrate Sender into OutgoingPipeline
OutgoingPipeline is renamed from OutgoingCallPipeline
