+++ src/main/java/de/rwth/idsg/steve/ocpp/ws/pipeline/IncomingPipeline.java
@@ -20,8 +20,8 @@ package de.rwth.idsg.steve.ocpp.ws.pipeline;
 
 import de.rwth.idsg.ocpp.jaxb.ResponseType;
 import de.rwth.idsg.steve.SteveException;
-import de.rwth.idsg.steve.ocpp.ws.data.CommunicationContext;
 import de.rwth.idsg.steve.ocpp.OcppVersion;
+import de.rwth.idsg.steve.ocpp.ws.data.CommunicationContext;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -44,15 +44,15 @@ public class IncomingPipeline {
 
     private final Serializer serializer = Serializer.INSTANCE;
 
-    private final Sender sender;
+    private final OutgoingPipeline outgoingPipeline;
     private final Deserializer deserializer;
     private final Map<OcppVersion, OcppCallHandler> handlerMap = new EnumMap<>(OcppVersion.class);
 
     @Autowired
-    public IncomingPipeline(Sender sender,
+    public IncomingPipeline(OutgoingPipeline outgoingPipeline,
                             Deserializer deserializer,
                             List<OcppCallHandler> handlers) {
-        this.sender = sender;
+        this.outgoingPipeline = outgoingPipeline;
         this.deserializer = deserializer;
         for (OcppCallHandler handler : handlers) {
             handlerMap.put(handler.getVersion(), handler);
@@ -67,7 +67,7 @@ public class IncomingPipeline {
         } catch (CommunicationContext.JsonCallParseException e) {
             var parseError = e.getParseError();
             var parseErrorStr = serializer.accept(parseError);
-            sender.accept(new CommunicationContext.Out(inMsg.route(), parseErrorStr));
+            outgoingPipeline.accept(new CommunicationContext.Out(inMsg.route(), parseErrorStr));
             return;
 
         } catch (CommunicationContext.JsonResponseInvalidException e) {
@@ -98,7 +98,7 @@ public class IncomingPipeline {
 
         var response = handler.accept(data);
         var responseStr = serializer.accept(response);
-        sender.accept(new CommunicationContext.Out(data.in().route(), responseStr));
+        outgoingPipeline.accept(new CommunicationContext.Out(data.in().route(), responseStr));
     }
 
     @SuppressWarnings("unchecked")
